Burned it to the ground...
Decided that refactoring wasn't enough, I ultimately deleted the project and started from scratch, this time truly building my own wrapper for Strava and Myfitnesspal's APIs. Also, switching the scope since I'm doing this to get some exposure to BigQuery, no more focus on the UI (I threw it out), instead the goal is for this code to be deployed to a VPC in the GCP, schedule it to run nightly, have it throw the data into the postgres db and then for the short term i'll schedule a dbt package to run nightly too but long term this will then go to BigQuery (I may skip Postgres entirely eventually) and then dbt will transform the data and finally I'll build something either in Power BI or R Shiny more than likely.